bot-service/requirements.txt
langchain-core==1.2.17
langchain-openai==1.1.10
pydantic==2.12.5
openai==2.26.0
python-dotenv==1.2.2
colorlog==6.10.1
fastapi==0.135.1
uvicorn==0.41.0
asyncpg==0.31.0
